src/backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
Pillow
django-cleanup
python-dotenv
djangorestframework-simplejwt